Renamed index.py to app.py, Deleted app.py. Removed print statements and old commented functions
